+++ src/modules/comment.js
@@ -1,5 +1,7 @@
 import { invAPI } from './api.js';
 
+/* API Functions */
+
 const commentURL = `${invAPI}comments?item_id=`;
 
 const getComment = async (id) => {
@@ -21,4 +23,37 @@ const addComment = async (id, user, desc) => {
   return data;
 };
 
-export { getComment, addComment };
+/* Display Functions */
+
+const newComment = document.querySelector('#add-comment');
+const commentForm = document.querySelector('#comment-form');
+const commentList = document.querySelector("#comment-list");
+
+const displayComment = async () => {
+  const commentData = await getComment("item5");
+  if (Object.keys(commentData).length = 0){
+    commentList.innerHTML = '';
+  } else {
+    commentData.forEach((item) => {
+        commentList.innerHTML+= `<li class="comment"> ${item.username}: ${item.comment} (${item.creation_date})`
+    })
+  };
+}
+
+newComment.addEventListener('click', (e) => {
+  e.preventDefault;
+  const commentor = document.querySelector('#username').value;
+  const comment = document.querySelector('#comment').value;
+  if (!commentor || !comment) {
+    e.preventDefault();
+    console.log("oops!");
+  } else {
+    addComment("item5", commentor, comment);
+    commentSection.innerHTML = '';
+    getComment();
+    displayComment();
+    commentForm.reset();
+  }
+});
+
+export { getComment, addComment , displayComment };
